pyproject.toml
[project]
name = "rag-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.4.0",
    "fastapi>=0.128.0",
    "google-genai>=1.60.0",
    "inngest>=0.5.15",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.7",
    "llama-index-core>=0.14.13",
    "llama-index-readers-file>=0.5.6",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "streamlit>=1.53.1",
    "tavily-python>=0.7.20",
    "uvicorn>=0.40.0",
]
